@@ -4,12 +4,16 @@
   "description": "simple JS class to manage color palettes",
   "main": "index.umd.js",
   "es2015": "index.js",
+  "type": "commonjs",
+  "module": "index.js",
   "scripts": {
-    "test": "jest",
+    "test": "npm run test:jest && npm run test:cli",
+    "test:jest": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "test:cli": "./generate.mjs -c ./test/colors.json -p ./test/palettes.json",
     "build": "rollup --config"
   },
   "bin": {
-    "paletterTo": "./generate.js"
+    "paletterTo": "./generate.mjs"
   },
   "repository": {
     "type": "git",
@@ -29,21 +33,23 @@
   },
   "homepage": "https://github.com/meodai/paletter#readme",
   "devDependencies": {
+    "@babel/core": "^7.23.2",
+    "@babel/preset-env": "^7.23.2",
     "@rollup/plugin-node-resolve": "^15.2.3",
-    "babel-core": "^6.26.0",
-    "babel-jest": "^21.2.0",
-    "babel-preset-latest": "^6.24.1",
-    "eslint": "^8.29.0",
+    "babel-jest": "^29.7.0",
+    "eslint": "^8.51.0",
     "eslint-config-google": "^0.9.1",
-    "jest": "^21.2.1",
+    "jest": "29.7",
     "rollup": "^4.1.4",
     "rollup-plugin-babel": "^3.0.2"
   },
   "jest": {
     "transform": {
-      "^.+\\.js?$": "babel-jest"
+      "^.+\\.js?$": "babel-jest",
+      "^.+\\.mjs$": "babel-jest"
     },
-    "collectCoverage": true
+    "collectCoverage": true,
+    "testMatch": ["**/__tests__/**/*.?(m)js?(x)", "**/?(*.)(spec|test).?(m)js?(x)"]
   },
   "dependencies": {
     "culori": "^3.2.0"
