@@ -1,7 +1,7 @@
-module.exports = function(obj, links, varPrefix = '', separator = '--') {
+export const toCSS = (obj, links, varPrefix = '', separator = '--') => {
   let CSSvars = ':root {\n';
   for (let palette in obj) {
-    let prefix = `--${varPrefix}${palette}`;
+    const prefix = `--${varPrefix}${palette}`;
     for (let key in obj[palette] ) {
       let color = obj[palette][key];
       const linkFromKey = links.find((c) => (c.from.key == `${palette}${separator}${key}`));
