@@ -1,4 +1,4 @@
-module.exports = (obj, connections, palette, palettes) => {
+export const toHTML = (obj, connections, palette, palettes) => {
   const css = `
   .styleguide__colors {
     display: -webkit-box;
