@@ -1,4 +1,4 @@
-module.exports = function (obj, links, varPrefix = '', separator = '--') {
+export const toSCSSvars = (obj, links, varPrefix = '', separator = '--') => {
   let SCSSVars = '';
   for (let palette in obj) {
     let prefix = `$${varPrefix}${palette}`;
