@@ -1,5 +1,5 @@
 /* eslint-disable guard-for-in */
-import { parse } from 'culori/fn';
+import { parse } from 'culori';
 
 /** Main paletter class */
 export default class Paletter {
