@@ -1,17 +1,13 @@
-import babel from 'rollup-plugin-babel';
 import { nodeResolve } from '@rollup/plugin-node-resolve';
 
 export default {
-  input: 'index.js',
+  input: 'index.mjs',
   output: {
     file: 'index.umd.js',
     format: 'umd',
     name: 'Paletter',
   },
   plugins: [
-    babel({
-      exclude: 'node_modules/**', // only transpile our source code
-    }),
     nodeResolve(),
   ],
 };
